Rename module: channels -> kawasemi
