+++ kawasemi/backends/hipchat.py
