+++ kawasemi/exceptions.py
