+++ tests/tests/test_init.py
@@ -6,20 +6,20 @@ try:
 except ImportError:
     import mock
 
-import channels
+import kawasemi
 
 
 class ChannelsModuleTestCase(TestCase):
     def test_load_module(self):
-        import channels.backends.slack
-        module_name = "channels.backends.slack"
-        module = channels._load_module(module_name)
-        self.assertEqual(module, channels.backends.slack)
+        import kawasemi.backends.slack
+        module_name = "kawasemi.backends.slack"
+        module = kawasemi._load_module(module_name)
+        self.assertEqual(module, kawasemi.backends.slack)
 
     def test_load_backend(self):
-        from channels.backends.slack import SlackChannel
-        klass_name = "channels.backends.slack.SlackChannel"
-        klass = channels._load_backend(klass_name)
+        from kawasemi.backends.slack import SlackChannel
+        klass_name = "kawasemi.backends.slack.SlackChannel"
+        klass = kawasemi._load_backend(klass_name)
         self.assertEqual(klass, SlackChannel)
 
     def test__send_channel_not_exist(self):
@@ -28,10 +28,10 @@ class ChannelsModuleTestCase(TestCase):
         }
 
         with self.assertRaises(Exception):
-            channels._send("message", settings, channel_name="not_exists",
+            kawasemi._send("message", settings, channel_name="not_exists",
                            fail_silently=False, options=None)
 
-    @mock.patch("channels._load_backend")
+    @mock.patch("kawasemi._load_backend")
     def test__send_all_channels(self, m):
         """test case for channel_name=None
 
@@ -40,24 +40,24 @@ class ChannelsModuleTestCase(TestCase):
         settings = {
             "CHANNELS": {
                 "channel1": {
-                    "_backend": "channels.backends.hipchat.HipChatChannel",
+                    "_backend": "kawasemi.backends.hipchat.HipChatChannel",
                     "api_id": "api_id",
                     "token": "token"
                 },
                 "channel2": {
-                    "_backend": "channels.backends.slack.SlackChannel",
+                    "_backend": "kawasemi.backends.slack.SlackChannel",
                     "url": "url"
                 }
             }
         }
 
-        channels._send("message", settings, channel_name=None,
+        kawasemi._send("message", settings, channel_name=None,
                        fail_silently=False, options=None)
         self.assertEqual(m.call_count, len(settings["CHANNELS"]))
         m.assert_any_call(settings["CHANNELS"]["channel1"]["_backend"])
         m.assert_any_call(settings["CHANNELS"]["channel2"]["_backend"])
 
-    @mock.patch("channels._load_backend")
+    @mock.patch("kawasemi._load_backend")
     def test__send_specific_channel(self, m):
         """test case for channel_name != None
 
@@ -66,17 +66,17 @@ class ChannelsModuleTestCase(TestCase):
         settings = {
             "CHANNELS": {
                 "channel1": {
-                    "_backend": "channels.backends.hipchat.HipChatChannel",
+                    "_backend": "kawasemi.backends.hipchat.HipChatChannel",
                     "api_id": "api_id",
                     "token": "token"
                 },
                 "channel2": {
-                    "_backend": "channels.backends.slack.SlackChannel",
+                    "_backend": "kawasemi.backends.slack.SlackChannel",
                     "url": "url"
                 }
             }
         }
 
-        channels._send("message", settings, channel_name="channel1",
+        kawasemi._send("message", settings, channel_name="channel1",
                        fail_silently=False, options=None)
         m.assert_called_once_with(settings["CHANNELS"]["channel1"]["_backend"])
