+++ tox.ini
@@ -28,7 +28,7 @@ deps =
     Django>=1.10,<1.11
 commands =
     pip install -e .[test]
-    coverage run --source channels/ ./runtests.py
+    coverage run --source kawasemi/ ./runtests.py
     coverage report
     coveralls
 
@@ -40,10 +40,10 @@ commands =
 
 [testenv:flake8]
 deps = flake8
-commands = flake8 channels/
+commands = flake8 kawasemi/
 
 [testenv:mypy]
 deps = mypy-lang
 commands =
-    mypy --py2 --silent-imports -m channels
-    mypy --silent-imports -m channels
+    mypy --py2 --silent-imports -m kawasemi
+    mypy --silent-imports -m kawasemi
